+++ lib/rubocop/remote_config.rb
@@ -61,7 +61,11 @@ module RuboCop
       when Net::HTTPRedirection
         request(URI.parse(response['location']), limit - 1, &block)
       else
-        response.error!
+        begin
+          response.error!
+        rescue StandardError => e
+          raise e, "#{e.message} while downloading remote config file #{uri}"
+        end
       end
     end